shipping: global single no-sharding company implementation
shipCompany_2.py::58 is having a problem: even though `ship` is
something like `Actor(Ship,xxxxxx)`, and manually inspecting the list
reveals an element of the same name, the check for whether the list
contains that `ship` fails. Furthermore, extracting that element from
the list and performing a == check with the `ship` gives False.

I'm a bit sad.