shipping: == on handle doesn't work
Use handleEQ() for the moment bc I don't know what else form should
it take.

This has the problem that we have to manually loop through
a list if we want to remove a handle from that list - we might
need to change the structure. Maybe a dict from shipname to handle?
We won't be allowing duplicated names then.